Changed Aggregation Calculations from float to decimal for precision. Geometric means Stack overflows now for large number tests, will swap out GM for R engine

+++ Noodles.Test/Utilities/PositionallyWeightedAggregation.cs
@@ -9,9 +9,9 @@ namespace Noodles.Test.Utilities
 {
     public class PositionallyWeightedAggregation : MeanCalculator
     {
-        protected override float InternalCalculate(IEnumerable<float> values)
+        protected override decimal InternalCalculate(IEnumerable<decimal> values)
         {
-            float result = 0;
+            decimal result = 0;
             int n = values.Count();
 
             values.Index((i, v) =>
